Rewrote imports, added threads, added send method
Imports with 'as' make writing code harder sometimes, I was struggling
with 'queue', so back to the long messy method for now.

Added verbosity to SLOBSClient, for future debug functionality.

Added threads dict for storing threads which I create.

Modified connect to return True, False. No need to make clients check.
It should return 'o' or give a fat error.

Modfiied auth to return both a bool and the payload, so you can figure
out what's wrong.

Added send, for sending a payload and setting up a callback, if wanted.
Going to move auth to using this.

Added recv, but not sure what for yet.